Moved my qth specifier to the beginning. Fixed rotator configuration code. Added disabled test code for the ax2ticks function.
Changed az2ticks function ti handle the fact that there is an overlap around the stop positions and to mode the antenna the shortest path to any position that overlaps,

When handle_track_az gets a QTH-locator instead of a numeric azimuth value, calculate the bearing to the center of the given square and ue the result for the azimuth. In this case that square will also be marked on the map.

Changed the data in msgq to contain a tuple of (what, data)

Changed the background process to us the what value in the queue do decide which message to send.

Push a send_origo containing message on connect.